Replace newlines with <br> in Message content
Message submissions contain newlines that people wish to have in their
message for the formatting. These newlines pass through the data
processing and are rendered into the HTML by minijinja.

However HTML does not care if you've got '\n' inside of a paragraph,
`<p>`, and renders those newlines as spaces. So we have to explicitly
replace the newlines with HTML line breaks `<br>` and use the `safe`
filter so minijinja doesn't escape the tags which would render the
`<br>` as literal text rather than HTML.

This does put the onus of sanitizing and validating the submitted
Messages on the data processing side.